seldonmodel/requirements.txt
pandas 
scikit-surprise 
scikit-learn 
numpy 
